requirements.txt
easyocr>=1.6.2
numpy>=1.21.0
Pillow>=9.0.0
torch>=1.13.0